requirements.txt
langchain
langchain-community
langchain-ollama
python-dotenv
faiss_cpu
flask
flask_cors